backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
langchain==0.3.13
langchain-groq==0.2.3
langchain-core==0.3.29
langgraph==0.2.60
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
httpx==0.28.1
python-multipart==0.0.20
